Renamed SOASLConnection to ASLConnection. The risk of any class name collison is minor. Refactored ivar names to reflect my current style.
Added -setSeverityFilterMask: method to SOLogger. Setting the filter mask affects the main thread ASLConnection, and the ASLConnection on the invocation thread if it's not main.
Modified behavior of -addFileDescriptor: and -removeFileDescriptor: to also change the main thread ASLConnection, as well as the ASLConnection on the invocation thread, if it's not main.

Added README.markdown in preparation for GitHub.